requirements.txt
langchain
langchain-groq
langgraph
chromadb
pydantic
pillow
requests
python-dotenv
